Add top comments, cleanup syntax and filenames
