+++ server/controllers/user.js
@@ -5,7 +5,7 @@
  */
 
 const jwt = require('../libraries/jsonWebToken');
-const User = require('../models/user');
+const User = require('../models/User');
 
 const signup = (req, res, next) => {
   const data = req.body;
