+++ server/index.js
@@ -1,3 +1,9 @@
+/**
+ * Server Entry Point.
+ *
+ * @module server/index
+ */
+
 const server = require('./libraries/server');
 const database = require('./libraries/database');
