+++ server/libraries/jsonWebToken.js
@@ -1,26 +1,28 @@
+/**
+ * JSON Web Token Library.
+ *
+ * @module server/libraries/jsonWebToken
+ */
+
 const jwt = require('jsonwebtoken');
 
 const ENV = require('../constants/env');
 
-/**
- * Create a token.
- */
+// Create a token.
 const sign = payload => new Promise((resolve, reject) => {
   jwt.sign(payload, ENV.JWT_SECRET, null, (error, token) => (
     error
-    ? reject(error)
-    : resolve(token)
+      ? reject(error)
+      : resolve(token)
   ));
 });
 
-/**
- * Verify a token.
- */
+// Verify a token.
 const verify = token => new Promise((resolve, reject) => {
   jwt.verify(token, ENV.JWT_SECRET, { complete: true }, (error, payload) => (
     error
-    ? reject(error)
-    : resolve(payload)
+      ? reject(error)
+      : resolve(payload)
   ));
 });
