+++ server/libraries/server.js
@@ -1,3 +1,9 @@
+/**
+ * Server Library.
+ *
+ * @module server/libraries/server
+ */
+
 const http = require('http');
 
 const app = require('./app');
@@ -6,22 +12,27 @@ const ENV = require('../constants/env');
 // Create HTTP server.
 const server = http.createServer(app);
 
+// Cache server address info.
+let address;
+
 // Log message when server starts.
 server.on('listening', () => {
-  console.log(`server running at port ${ENV.PORT}`);
+  address = server.address()
+
+  console.log(`server running at port ${address.port}`);
 });
 
 // Log message when server stops.
 server.on('close', () => {
-  console.log(`server stopped at port ${ENV.PORT}`);
+  console.log(`server stopped at port ${address.port}`);
 });
 
 // Start HTTP server.
 const start = (port = ENV.PORT, host = ENV.HOST) => new Promise((resolve, reject) => {
   server.listen(port, host, err => (
     err
-    ? reject(err)
-    : resolve()
+      ? reject(err)
+      : resolve()
   ));
 });
 
@@ -29,8 +40,8 @@ const start = (port = ENV.PORT, host = ENV.HOST) => new Promise((resolve, reject
 const stop = () => new Promise((resolve, reject) => {
   server.close(err => (
     err
-    ? reject(err)
-    : resolve()
+      ? reject(err)
+      : resolve()
   ));
 });
