+++ server/routes/web.js
@@ -1,10 +1,15 @@
+/**
+ * Web Routes.
+ *
+ * @module server/routes/web
+ */
+
 const { Router } = require('express');
 
 const staticController = require('../controllers/static');
 
 const webRouter = Router();
 
-// Render react app.
 webRouter.get('*', staticController.index);
 
 module.exports = webRouter;
